pyproject.toml
[project]
name = "kobo-quotes-to-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ell-ai[sqlite,studio]>=0.0.17",
    "rich>=14.0.0",
    "typer>=0.16.0",
]
